refactor: Move wifi-only policy use case to combined package
- Move `ApplyWifiOnlyPolicyUseCase` to `combined` package
- Replace `DownloadUseCase` with explicit download-related use cases
- Update `SettingsViewModel` import to new package
